Initial commit
The intent of this branch is to start the refactoring process by
moving the Resolver class out of __init__.py and into it's own
module file.  The biggest issue with this move was fixing the import
issues that arose from the move.

The secondary changes involved some cleanup removing __future__ calls
needed to support Python2.  There are no longer needed since Hab is
now Python3 only.

Finally, I cleaned up some f-string errors when using quotes within
the f-string itself.  I fixed these by using the `!r` conversion flag
which calls the __rep__() method.  The result is the variable in the
{} will be wrapped in quotes avoiding the need to use extra quotes in
the f-string.

i.e. `f"{key_upper!r} is a reserved environment variable"`
